Compled Weapon system overhawl. Replaced room icon/stat UI placeholder sprites. Added Ally Knight's weapon sprite. Bug Fixes
Added gear tab QOL - Selecting an empty main gear slot, auto opens owned gear.

+++ Assets/Scripts/HitAreaManager.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class HitAreaManager : MonoBehaviour
 {
     [SerializeField] private Transform hitAreasTrans;
-    [SerializeField] private int minYPos;
-    [SerializeField] private int maxYPos;
+    [SerializeField] private float minYPos;
+    [SerializeField] private float maxYPos;
 
     public void UpdateHitAreaPos()
     {
